feat(domain): Strategy refund family + iter 2 state transitions + e-Ticket
- RefundPolicy interface (BigDecimal) + Full/Partial/No concretes
- FareRule.checkRefundPolicy mirrors resolvePolicy mapping
- Refund / RefundRequest constructors + state transitions
- State pattern transition bodies for iter 2:
  Confirmed.issueTicket → Ticketed (creates Ticket per passenger)
  Confirmed/Ticketed.requestCancellation → CancellationRequested
  CancellationRequested.confirmCancellation → Cancelled
  Cancelled.requestRefund → RefundRequested (with FareRule check)
  RefundRequested.processRefundDecision(approved) → Refunded/Cancelled
- Ticket.generate(reservation, passenger, seat) factory (KE-yyMMdd-NNNN)
- Reservation.findByPnr registry + addTicket helper
- Segment setter to wire FlightSchedule into Itinerary

+++ src/com/koreanair/reservation/domain/payment/FullRefundPolicy.java
@@ -1,9 +1,11 @@
 package com.koreanair.reservation.domain.payment;
 
+import java.math.BigDecimal;
+
 public class FullRefundPolicy implements RefundPolicy {
 
     @Override
-    public int calculateRefundAmount(int baseAmount) {
+    public BigDecimal calculateRefundAmount(BigDecimal baseAmount) {
         return baseAmount;
     }
